requirements.txt
# Django y REST Framework
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
django-extensions==4.1

# Autenticación y seguridad
PyJWT==2.9.0

# Base de datos
sqlparse==0.5.3
psycopg2-binary==2.9.9  # Adaptador PostgreSQL para Django

# Procesamiento de datos y ML
numpy==2.2.5
pandas==2.2.3
scikit-learn==1.7.0
scipy==1.16.0
bottleneck==1.3.7

# Visualización
matplotlib==3.10.3
seaborn==0.13.2
contourpy==1.3.2
cycler==0.12.1

# Utilidades
setuptools==72.1.0
six==1.16.0
threadpoolctl==3.6.0
tzdata==2023.3
wheel==0.45.1

# Excel y reportes
xlrd==2.0.1
openpyxl==3.1.2
reportlab==4.1.0

# Cache y optimización
redis==5.0.1

# Monitoreo y logging
sentry-sdk==2.0.0

# Testing
pytest==8.0.0
pytest-django==4.8.0